refactor: centralize isolation level validation with static supportedIsolationLevels per driver
- Add IsolationLevels const array and derive IsolationLevel type from it
- Add static readonly supportedIsolationLevels to all 11 driver classes
- Add standalone validateIsolationLevel() in driver/validate-isolation-level.ts
- Replace manual if/throw checks in all query runners with validateIsolationLevel()
- Add DataSource-level validation in SqlServerDriver.createPool()
- Rename test folder database-specific-isolation → isolation-level
- Restructure tests: defined in data source / defined for transaction / supported / unsupported
- Add spanner isolation test, add sqljs to sqlite test
- Tests derive supported/unsupported levels from driver static properties
