+++ src/driver/types/IsolationLevel.ts
@@ -1,6 +1,9 @@
-export type IsolationLevel =
-    | "READ UNCOMMITTED"
-    | "READ COMMITTED"
-    | "REPEATABLE READ"
-    | "SERIALIZABLE"
-    | "SNAPSHOT" // SQL Server specific isolation level
+export const IsolationLevels = [
+    "READ COMMITTED",
+    "READ UNCOMMITTED",
+    "REPEATABLE READ",
+    "SERIALIZABLE",
+    "SNAPSHOT",
+] as const
+
+export type IsolationLevel = (typeof IsolationLevels)[number]
